+++ public/sw.js
@@ -1,4 +1,4 @@
-const OED_VERSION = '4.0.20200806.4';
+const OED_VERSION = '4.0.20200806.5';
 const OED_BASE = 'OED';
 const ACE_VERSION = '1.4.12';
 const ACE_BASE = 'Ace';
@@ -88,7 +88,6 @@ const OED_URLS = [
     './statusbar.js',
     './sw.js',
     './theme_helper.js',
-    './toggle_drawer_event.js',
     './ui_helper.js',
 ];
 const ACE_URLS = [