+++ public/toggle_drawer_event.js
@@ -1,5 +1,0 @@
-export class ToggleDrawerEvent extends CustomEvent {
-    constructor() {
-        super('Drawer:toggle');
-    }
-}